requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow-cpu>=2.20.0
numpy>=1.24.3
scikit-learn>=1.3.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
